+++ soc/xtensa/intel_adsp/common/include/adsp-vectors.h
